projects/resume_analyzer/requirements.txt
streamlit
langgraph
langchain-openai
python-dotenv
pypdf
python-docx
